feat(runner-schema): extract v0 schema data structures into assay-runner-schema crate (Phase 2D Slice 1)
Phase 2D Slice 1 of the Assay-Runner extraction roadmap. Move the
data-structure half of the assay.runner.*.v0 contracts out of
crates/assay-runner-spike/ into a new publish-disabled crate
crates/assay-runner-schema/.

## Extraction boundary impact

This PR is **type A** per the extraction-roadmap.md per-PR discipline
rule: it resolves named structural blockers.

- Blocker 1 (assay.runner.*.v0 schemas live in
  crates/assay-runner-spike/src/) — RESOLVED FULLY. All five schema
  data-structure modules now live in crates/assay-runner-schema/.
- Blocker 2 (archive verification crosses an unresolved boundary
  conflict in archive.rs) — RESOLVED PARTIALLY. The manifest semantics
  half (ARCHIVE_MANIFEST_SCHEMA, path constants, ArchiveFile,
  ArchiveManifest) moved to assay-runner-schema. The assembly half
  (RunnerSpikeArchive, RunnerSpikeArchiveError, write, archive_files,
  build_manifest) stays in assay-runner-spike and moves to
  assay-runner-core at Slice 2 per the roadmap.

## What moved to crates/assay-runner-schema/

Five modules + lib.rs:

- health.rs (whole file) - ObservationHealth, status enums,
  ObservationHealthError, OBSERVATION_HEALTH_SCHEMA, 8 unit tests
- surface.rs (whole file) - CapabilitySurface,
  CapabilitySurfaceError, CAPABILITY_SURFACE_SCHEMA, 4 unit tests
- correlation.rs (whole file) - CorrelationReport, CorrelationStatus,
  BindingWindow, CorrelationBinding, CorrelationReportError,
  CORRELATION_REPORT_SCHEMA, 3 unit tests
- sdk_event.rs (subset of spike/sdk.rs) - SDK_EVENT_SCHEMA,
  SdkLayerEvent. Parsing logic (SdkLayerCapture, SdkLayerError,
  from_sdk_ndjson) stays in spike.
- archive_manifest.rs (subset of spike/archive.rs) -
  ARCHIVE_MANIFEST_SCHEMA, all path constants, ArchiveFile,
  ArchiveManifest. Assembly logic (RunnerSpikeArchive, write) stays in
  spike.

New crate dependencies (stdlib + serde + thiserror only):

- serde (with derive + std)
- thiserror
- serde_json (dev only, for surface tests)

Explicitly NOT included (forbidden by roadmap):

- assay-common, assay-monitor, assay-cli, assay-ebpf
- flate2, tar, sha2, hex, uuid (those are archive assembly deps)

publish = false until Slice 7 per the roadmap.

## Compatibility

crates/assay-runner-spike/src/lib.rs now re-exports the moved types
via `pub use assay_runner_schema::{...}`. External consumers
(assay-cli, scripts/ci/runner-spike-*.sh, tests/fixtures) continue to
import via `assay_runner_spike::Type` unchanged.

Within the spike crate, internal references like
`crate::health::ObservationHealthError` resolve through the re-export
or have been migrated to `assay_runner_schema::ObservationHealthError`
where the explicit crate path is clearer (archive.rs, kernel.rs,
policy.rs).

## Lane-check classifier updated

crates/assay-runner-schema/ added to the all_prefixes list in
scripts/ci/assay_runner_lane_check.py: schema changes are
runner-impacting because they affect contract semantics that
delegated acceptance asserts. Two new self-test cases pin this
behaviour against future classifier refactors:

- crates/assay-runner-schema/src/lib.rs -> Gate.ALL
- crates/assay-runner-schema/Cargo.toml -> Gate.ALL

## Boundary-map ownership table updated

docs/reference/runner/boundary-map.md § "Current In-Repo Ownership"
now lists crates/assay-runner-schema/ as a separate row with the five
schema files. The old assay-runner-spike/src/{health,surface,
correlation,sdk,archive}.rs rows have been updated to reflect that
the schema half moved.

docs/reference/runner/boundary-map.md § "Active Boundary Conflicts"
now marks the archive.rs conflict as "Partially resolved by Phase 2D
Slice 1" with explicit pointer to Slice 2 for the assembly half.

## Readiness checkpoint updated

docs/reference/runner/platform-and-extraction-readiness.md §
"Extraction Readiness" now records blocker 1 as resolved and
blocker 2 as partially resolved by this slice, with pointers to
Slice 2 for the remaining work.

## What this PR explicitly does not do

- modify any v0 schema string value, field set, or value vocabulary
- modify any v0 contract document semantically (only ownership-table
  updates that reflect the structural change)
- introduce new fixtures, workflows, or CI lane rules
- introduce new dependencies (schema crate uses serde + thiserror only)
- open Slice 2 (assay-runner-core crate); that is a separate future PR

## Validation (local)

- cargo build -p assay-runner-schema: ok
- cargo test -p assay-runner-schema --lib: 14 passed (8 health, 4
  surface, 3 correlation; archive_manifest + sdk_event are pure
  data types with no behaviour to test in this slice)
- cargo build -p assay-runner-spike: ok
- cargo test -p assay-runner-spike --lib: 49 passed (all previously
  green tests continue to pass through the re-export surface)
- cargo build -p assay-cli: ok (external consumer compiles unchanged)
- cargo clippy -p assay-runner-schema -p assay-runner-spike
  -p assay-cli --all-targets -- -D warnings: ok
- python3 scripts/ci/assay_runner_lane_check.py --self-test: ok
- mkdocs build --strict: ok, no new warnings
- pre-commit hooks: ok (fmt, typos, merge conflicts, trailing
  whitespace, secret hygiene)
- pre-push hooks: ok (clippy on changed crates, linux compile gate)

A delegated gates=all run is required for this PR per the lane-check
classifier (Cargo.toml and crates/assay-runner-spike/ changes both
require Gate.ALL).
