+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
   "crates/assay-ebpf", "crates/assay-common",
   "assay-python-sdk", "crates/assay-policy", "crates/assay-evidence", "crates/assay-sim",
   "crates/assay-registry",
+  "crates/assay-runner-schema",
   "crates/assay-runner-spike",
 ]
 default-members = [
@@ -80,6 +81,7 @@ assay-mcp-server = { path = "crates/assay-mcp-server", version = "3.10.2" }
 assay-evidence = { path = "crates/assay-evidence", version = "3.10.2" }
 assay-sim = { path = "crates/assay-sim", version = "3.10.2" }
 assay-registry = { path = "crates/assay-registry", version = "3.10.2" }
+assay-runner-schema = { path = "crates/assay-runner-schema", version = "3.10.2" }
 assay-runner-spike = { path = "crates/assay-runner-spike", version = "3.10.2" }
 
 # Common dependencies
