+++ crates/assay-runner-schema/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "assay-runner-schema"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+description = "Versioned schema types and constants for the Assay-Runner v0 contracts (Phase 2D Slice 1)."
+readme.workspace = true
+publish = false
+
+[dependencies]
+serde = { workspace = true, features = ["derive", "std"] }
+thiserror = { workspace = true }
+
+[dev-dependencies]
+serde_json = { workspace = true, features = ["std"] }
+
+[lints]
+workspace = true
