+++ crates/assay-runner-schema/src/health.rs
