+++ crates/assay-runner-spike/Cargo.toml
@@ -12,6 +12,7 @@ publish = false
 [dependencies]
 assay-common = { workspace = true, features = ["std"] }
 assay-monitor = { workspace = true }
+assay-runner-schema = { workspace = true }
 serde = { workspace = true, features = ["derive", "std"] }
 serde_json = { workspace = true, features = ["std"] }
 sha2 = { workspace = true }
