+++ crates/assay-runner-spike/src/sdk.rs
@@ -1,23 +1,8 @@
 use crate::{run::is_safe_run_id, RunnerSpikeArchive, SdkLayerStatus};
-use serde::{Deserialize, Serialize};
+use assay_runner_schema::{SdkLayerEvent, SDK_EVENT_SCHEMA};
 use serde_json::Value;
 use std::collections::BTreeSet;
 use thiserror::Error;
-
-pub const SDK_EVENT_SCHEMA: &str = "assay.runner.sdk_event.v0";
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SdkLayerEvent {
-    pub schema: String,
-    pub run_id: String,
-    pub seq: u64,
-    pub event_type: String,
-    pub source: String,
-    pub sdk_name: Option<String>,
-    pub sdk_version: Option<String>,
-    pub tool_call_id: Option<String>,
-    pub tool: Option<String>,
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct SdkLayerCapture {
